Updated gameplay loop
-Player can select item to sell
-slider to select initial price
- buttons to increase and decrease price.

+++ Assets/Scripts/BaseCharacter.cs
@@ -21,7 +21,7 @@ public class BaseCharacter : ScriptableObject
     [SerializeField] public int machineryDesire;
     [SerializeField] public int luxuryDesire;
     [SerializeField] public TextDialogue[] introText;
-    [SerializeField] public float patiece;
+    [SerializeField] public int patiece;
 }
 
 public class Script : MonoBehaviour
